ecs refactoring
