+++ src/scene/components/CameraComponent.h
@@ -1,0 +1,12 @@
+#ifndef RPG_CAMERACOMPONENT_H
+#define RPG_CAMERACOMPONENT_H
+
+#include "../../graphics/Camera2D.h"
+
+// TODO: переделать компонент камеры
+struct CameraComponent
+{
+    Camera2D *camera;
+};
+
+#endif //RPG_CAMERACOMPONENT_H
