+++ src/scene/components/NameComponent.h
@@ -1,0 +1,9 @@
+#ifndef RPG_NAMECOMPONENT_H
+#define RPG_NAMECOMPONENT_H
+
+struct NameComponent
+{
+    std::string name;
+};
+
+#endif //RPG_NAMECOMPONENT_H
