+++ src/scene/components/SpriteRendererComponent.cpp
@@ -1,0 +1,5 @@
+#include "SpriteRendererComponent.h"
+
+SpriteRendererComponent::SpriteRendererComponent(Texture texture)
+        : texture(texture),
+          textureRect(0, 0, texture.getWidth(), texture.getHeight()) {}
