+++ src/scene/components/TextRendererComponent.cpp
@@ -1,0 +1,4 @@
+#include "TextRendererComponent.h"
+
+TextRendererComponent::TextRendererComponent(Font *font, std::string text)
+        : font(font), text(text) {}
