+++ src/scene/components/TextRendererComponent.h
@@ -1,0 +1,27 @@
+#ifndef RPG_TEXTRENDERERCOMPONENT_H
+#define RPG_TEXTRENDERERCOMPONENT_H
+
+#include "../../graphics/Font.h"
+
+enum class HorizontalAlign
+{
+    Left, Right, Center
+};
+
+enum class VerticalAlign
+{
+    Top, Bottom, Center
+};
+
+struct TextRendererComponent
+{
+    Font *font;
+    std::string text;
+    glm::vec4 color{1.f};
+    HorizontalAlign horizontalAlign = HorizontalAlign::Left;
+    VerticalAlign verticalAlign = VerticalAlign::Bottom;
+
+    TextRendererComponent(Font *font, std::string text);
+};
+
+#endif //RPG_TEXTRENDERERCOMPONENT_H
